comments that survived what they described, and exports nobody wanted
core/sys/main.h still described main_api, which was deleted last pass and
by then existed nowhere else in the tree -- the comment was the last
reference to it. core/main.h claimed a machine with a real bus "takes the
two as one call and routes on the device number", which the RIA has never
done: it calls device 0 directly and puts device 1 on the wire.

Two comments outlived their declarations by one commit each and had
slid onto whatever followed: one describing the deleted vga_prog_mode
now sat above vga_prog_valid, one describing the deleted msc_api_chdir
above the file driver. And prog.h said vga_prog_valid was "shared even by
a machine that keeps no table" while the Pocket re-implements it twelve
lines away -- the same header says the truth thirty lines up.

Five include guards still named core/sys, where their files have not
lived for two passes. kbd_text.h's is left alone on purpose: that file is
going back to core/sys, and its guard is right again when it does.

Four things were exported to nobody. pix_deliver and dir_push_filinfo
were seams until last pass folded both sides into one file, and their
only callers are now in that same file. msc_dt never had an outside
caller. sw/pro.h re-declared pro_set_exit_code that the Pocket already
gets from core/api/pro.h.

And vga_vsync_scanline returned int in one header and int16_t in the
other, which nothing caught because no translation unit includes both.
It returns what it actually returns now, matching the rest of the video
API.

Firmware unchanged on both.

+++ src/core/vga/vga.c
@@ -98,7 +98,7 @@ void vga_task(void)
     }
 }
 
-int vga_vsync_scanline(void)
+int16_t vga_vsync_scanline(void)
 {
     /* Mirror the firmware (vga_scanline_complete): vsync fires at the highest
      * scanline any program renders, clamped to / falling back to the canvas
